application/backend/requirements.txt
apturl==0.5.2
asn1crypto==0.24.0
backports.entry-points-selectable==1.1.0
Brlapi==0.6.6
certifi==2021.5.30
chardet==3.0.4
click==8.0.1
colorama==0.3.7
command-not-found==0.3
cryptography==2.1.4
cupshelpers==1.0
dataclasses==0.8
defer==1.0.6
distlib==0.3.3
distro-info===0.18ubuntu0.18.04.1
filelock==3.2.0
Flask==2.0.1
Flask-SQLAlchemy==2.5.1
greenlet==1.1.2
httplib2==0.9.2
idna==2.6
importlib-metadata==4.8.1
importlib-resources==5.2.2
itsdangerous==2.0.1
Jinja2==3.0.1
keyring==10.6.0
keyrings.alt==3.0
language-selector==0.1
launchpadlib==1.10.6
lazr.restfulclient==0.13.5
lazr.uri==1.0.3
louis==3.5.0
macaroonbakery==1.1.3
Mako==1.0.7
MarkupSafe==2.0.1
mysql-connector==2.2.9
netifaces==0.10.4
oauth==1.0.1
olefile==0.45.1
pexpect==4.2.1
Pillow==5.1.0
pipenv==2021.5.29
platformdirs==2.4.0
protobuf==3.0.0
pycairo==1.16.2
pycrypto==2.6.1
pycups==1.9.73
PyGObject==3.26.1
pymacaroons==0.13.0
PyNaCl==1.1.2
pyRFC3339==1.0
python-apt==1.6.5+ubuntu0.7
python-dateutil==2.6.1
python-debian==0.1.32
python-dotenv==0.19.0
pytz==2018.3
pyxdg==0.25
PyYAML==3.12
reportlab==3.4.0
requests==2.18.4
requests-unixsocket==0.1.5
SecretStorage==2.3.1
simplejson==3.13.2
six==1.16.0
SQLAlchemy==1.4.25
system-service==0.3
systemd-python==234
typing-extensions==3.10.0.2
ubuntu-advantage-tools==27.2
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
urllib3==1.22
usb-creator==0.3.3
virtualenv==20.8.1
virtualenv-clone==0.5.7
wadllib==1.3.2
Werkzeug==2.0.1
xkit==0.0.0
zipp==3.6.0
zope.interface==4.3.2
